requirements.txt
altgraph==0.16.1
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.0
crypto==1.4.1
future==0.16.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
macholib==1.11
mccabe==0.6.1
Naked==0.1.31
pefile==2018.8.8
Pillow==5.2.0
PyInstaller==3.4
pylint==2.1.1
pywin32-ctypes==0.2.0
PyYAML==3.13
requests==2.19.1
shellescape==3.4.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
